requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.1.0
certifi==2023.11.17
environs==10.0.0
frozenlist==1.4.1
idna==3.6
magic-filter==1.0.12
marshmallow==3.20.1
multidict==6.0.4
packaging==23.2
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
typing_extensions==4.9.0
yarl==1.9.4
